prime/requirements.txt
# Alfred Prime - Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# gRPC
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2

# Claude API
anthropic==0.18.1

# Telegram
python-telegram-bot==20.8

# HTTP client
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
